+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -3906,7 +3906,7 @@ public class ServiceHost implements ServiceRequestSender {
 
     private static void failRequest(Operation request, int statusCode, int errorCode, Throwable e) {
         request.setStatusCode(statusCode);
-        ServiceErrorResponse r = Utils.toServiceErrorResponse(e, request);
+        ServiceErrorResponse r = Utils.toServiceErrorResponse(e);
         r.statusCode = statusCode;
         r.errorCode = errorCode;