+++ CHANGELOG.md
@@ -22,6 +22,11 @@
   Documents which match the query but have the groupBy term missing are returned
   under a special group "DocumentsWithoutResults".
 
+* Remove Utils.toServiceErrorResponse(Throwable t, Operation op).
+  Utils.toServiceErrorResponse(Throwable t) or
+  toValidationErrorResponse(Throwable t, Operation op) can be used. The
+  Operation argument is needed for localization of the error.
+
 ## 1.3.5
 
 * Add a new GatewayService to facilitate with blue/green upgrades of xenon